third_party/blink/web_tests/fast/dom/Document/readystate-expected.txt

Tests that the document's readyState is set properly at various phases during load.
PASS document.readyState is "loading"
PASS document.onreadystatechange is null
PASS document.readyState is "loading"
PASS subdocument.readyState is "loading"
PASS subdocument.readyState is "complete"
PASS htmlDocument.readyState is "loading"
PASS dynamicIframe.contentDocument.readyState is "complete"
PASS dynamicIframe.contentDocument.onreadystatechange is null
PASS dynamicIframe.contentDocument.readyState is "loading"
PASS dynamicIframe.contentDocument.readyState is expectedDynamicStates[nextExpectedDynamicState]
PASS dynamicIframe.contentDocument.readyState is expectedDynamicStates[nextExpectedDynamicState]
PASS dynamicIframe.contentDocument.readyState is "complete"
PASS event.target.readyState is expectedStates[nextExpectedState]
PASS document.readyState is "interactive"
PASS readyStateChangeFired is true
PASS document.readyState is "interactive"
PASS document.readyState is "interactive"
PASS event.target.readyState is expectedStates[nextExpectedState]
PASS nextExpectedState is expectedStates.length
PASS nextExpectedDynamicState is expectedDynamicStates.length
PASS document.readyState is "complete"
PASS successfullyParsed is true

TEST COMPLETE
